finish off the todos
